完成博客标签接口  修改注释信息  rsw

+++ blog-system/src/main/java/com/blog/module/business/service/impl/BlogTypeServiceImpl.java
@@ -1,5 +1,7 @@
 package com.blog.module.business.service.impl;
 
+import cn.hutool.core.collection.CollectionUtil;
+import com.blog.exception.BadRequestException;
 import com.blog.module.business.domain.BlogType;
 import com.blog.module.business.mapper.BlogTypeMapper;
 import com.blog.module.business.service.BlogTypeService;
@@ -8,8 +10,11 @@ import com.blog.page.vo.PageVo;
 import com.github.pagehelper.Page;
 import com.github.pagehelper.PageHelper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.util.CollectionUtils;
 
+import java.util.Collections;
 import java.util.List;
 
 /**
@@ -33,7 +38,10 @@ public class BlogTypeServiceImpl implements BlogTypeService {
      */
     @Override
     public void saveBlogType ( BlogType blogType ) {
-        blogTypeMapper.insert(blogType);
+        int count = blogTypeMapper.insertSelective(blogType);
+        if (count == 0) {
+            throw new BadRequestException("新增博客类型失败！");
+        }
     }
 
     /**
@@ -45,7 +53,10 @@ public class BlogTypeServiceImpl implements BlogTypeService {
      */
     @Override
     public void deleteBlogType ( Long blogTypeId ) {
-        blogTypeMapper.deleteByPrimaryKey(blogTypeId);
+        int count = blogTypeMapper.deleteByPrimaryKey(blogTypeId);
+        if (count == 0) {
+            throw new BadRequestException("删除博客类型失败！");
+        }
     }
 
     /**
@@ -57,7 +68,13 @@ public class BlogTypeServiceImpl implements BlogTypeService {
      */
     @Override
     public void deleteBlogTypes ( List<Long> blogTypeIds ) {
-        blogTypeMapper.deleteByIdList(blogTypeIds);
+        if (CollectionUtil.isEmpty(blogTypeIds)) {
+            throw new BadRequestException(HttpStatus.NOT_FOUND, "传入所需要删除的ID集合不能为空");
+        }
+        int count = blogTypeMapper.deleteByIdList(blogTypeIds);
+        if (count == 0) {
+            throw new BadRequestException("删除博客类型失败！");
+        }
     }
 
     /**
@@ -69,7 +86,10 @@ public class BlogTypeServiceImpl implements BlogTypeService {
      */
     @Override
     public void updateBlogType ( BlogType blogType ) {
-        blogTypeMapper.updateByPrimaryKeySelective(blogType);
+        int count = blogTypeMapper.updateByPrimaryKeySelective(blogType);
+        if (count == 0) {
+            throw new BadRequestException("修改博客类型失败！");
+        }
     }
 
     /**
@@ -81,8 +101,12 @@ public class BlogTypeServiceImpl implements BlogTypeService {
      * Date: 2020/4/13 9:16
      */
     @Override
-    public BlogType queryBlogById ( Long blogTypeId ) {
-        return blogTypeMapper.selectByPrimaryKey(blogTypeId);
+    public BlogType queryBlogTypeById ( Long blogTypeId ) {
+        BlogType blogType = blogTypeMapper.selectByPrimaryKey(blogTypeId);
+        if (blogType == null) {
+            throw new BadRequestException("暂无此博客类型！");
+        }
+        return blogType;
     }
 
     /**
@@ -94,12 +118,15 @@ public class BlogTypeServiceImpl implements BlogTypeService {
      * Date: 2020/4/13 9:16
      */
     @Override
-    public PageResultDto<BlogType> queryBlogAll ( PageVo pageVo ) {
+    public PageResultDto<BlogType> queryBlogTypeAll ( PageVo pageVo ) {
         if (pageVo != null) {
             PageHelper.startPage(pageVo.getCurrentPage(), pageVo.getRows(), pageVo.getSort());
         }
         //查询结果转换为Page对象
         Page<BlogType> blogTypePage = (Page<BlogType>) blogTypeMapper.selectAll();
+        if (CollectionUtils.isEmpty(blogTypePage)) {
+            throw new BadRequestException("暂无博客类型！");
+        }
         return new PageResultDto<>(blogTypePage.getTotal(), blogTypePage.getPageSize(), blogTypePage);
     }
 }
